project/assign_app/analyzer.py
from google.cloud import vision 



def analyze_image(image_path):
    client = vision.ImageAnnotatorClient()

    with open(image_path, 'rb') as image_file:
        content = image_file.read()

    image = vision.Image(content=content)
    response = client.label_detection(image=image)
    labels = response.label_annotations

    results = []
    for label in labels:
        results.append(label.description)

    return results